[REF] Board 초기화를 인터페이스로 분리
